Second commit. App is now 100% clean using C#, HTML, and jQuery.

+++ ToDoWebsite/ToDo/ToDoListEndpoint.cs
@@ -42,18 +42,42 @@ namespace ToDoWebsite.ToDo
             return FubuContinuation.RedirectTo<ToDoListInputModel>();
         }
 
-        public FubuContinuation post_Completed(ToDoListPostInputModel inputModel)
+        public FubuContinuation post_Removed(ToDoListPostInputModel inputModel)
         {
-            var delete = inputModel.Item;
+                var left = inputModel.Item;
+           
+                var tempFile = Path.GetTempFileName();
+
+                var linesToKeep = File.ReadLines(path).Where(l => l != left);
+
+                File.WriteAllLines(tempFile, linesToKeep);
+
+                File.Delete(path);
+                File.Move(tempFile, path);
             
+           
+
+            return FubuContinuation.RedirectTo<ToDoListInputModel>();
+        }
+
+        public FubuContinuation post_Edited(ToDoListPostInputModel inputModel)
+        {
+            var left = inputModel.Item;
+            var right = inputModel.Mod;
+
             var tempFile = Path.GetTempFileName();
-            var linesToKeep = File.ReadLines(path).Where(l => l != delete);
+            if (left != right)
+            {
 
-            File.WriteAllLines(tempFile, linesToKeep);
+                var linesToReplace = File.ReadAllLines(path).ToList();
+                var index = linesToReplace.IndexOf(left);
+                linesToReplace.Remove(left);
+                linesToReplace.Insert(index, right);
+                File.WriteAllLines(tempFile, linesToReplace);
 
-            File.Delete(path);
-            File.Move(tempFile, path);
-            
+                File.Delete(path);
+                File.Move(tempFile, path);
+            }
             return FubuContinuation.RedirectTo<ToDoListInputModel>();
         }
 
@@ -71,6 +95,8 @@ namespace ToDoWebsite.ToDo
      public class ToDoListPostInputModel
     {
         public string Item { get; set; }
+        public string Mod { get; set; }
+
     }
 
 }
